Fixing localization issues with the server scripts.
server/run_server_locally.sh - Changed some of the settings to make it
work a little nicer with the external version of GAE.  Using the
defaults for --auth_domain (since most don't have @google addresses)
and --address (since most prefer localhost).  There is also a more
involved way to run the development server if you need to ensure
presistant data in the development datastore.  Documented why one may
want to use the -c (clean) option.

server/set_version.sh - chmod 755

server/update_dev_server.sh - Changed some of the internal Google
settings to be more exteranly friendly.  Since this is open source the
app ID may change and was factored out.  The user email shouldn't
necesarily be the local user name and in many cases will not be a
@google address.  Since most will use their gmail/google email address
with their git config most of the time it is safe to pull from there.
Also the domain (for now) should be ___.appspot.com instead of
google.com.

There is more localization issues that need to be fixed in the base.html
template to remove Moma "goto" links and Bazaar references.