Responding to feedback for commit 365a6bbb5bf8.
Fixed typo in upload_local_data.sh, removed line from
download_dev_data.sh that made a backup copy of the downloaded data, and
gave the user warning that we would be writing to a few files.  This is
also an attempt to try rebase instead of pull from master.  The former
seems to be a bit noisy.