Be more aggressive about persisting measurement results and stats.
Prior to this change, measurement results were only being persisted on
MeasurementScheduler.onDestroy() which I believe is not guaranteed to
be called on a low-memory eviction. With this change,
MeasurementScheduler now persists measurement results frequently.

Note that MeasurementScheduler's pendingTasks data structure is still
kept in memory and can be lost on eviction, meaning that completed
measurement tasks can be lost before being uploaded on checkin.

Also fixed up a few minor GUI bugs and renamed a couple of functions
with typos.